requirements.txt
discord==1.7.3
python-dotenv==0.19.0
pysqlcipher3==1.0.4
asyncpg==0.24.0
yagmail==0.14.260
imap-tools==0.46.0
pysqlcipher3==1.0.4
keyring==23.2.1
